nimiGameランダム落下、あたり判定、スコア加算

+++ Yanai/taro/nimiPlayer.h
@@ -4,8 +4,8 @@
 #include "lib/appEnv.hpp"
 
 struct Player{
-	float x, y;
-	float width, height;
+	int x, y;
+	int width, height;
 };
 
 Player player{
